Add mobile responsive manual payment entry screen
