+++ tailwind.config.js
@@ -209,6 +209,7 @@ module.exports = {
         129: "33rem",
         138: "34.5rem",
         160: "40rem",
+        "1/5": "20%"
       },
       padding: {
         "36/100": "36.66666%",
